pyproject.toml
[project]
name = "project-registry"
version = "0.1.0"
description = "Project and time tracking CLI for motion design and VFX work"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "waitress>=3.0.2",
    "pyyaml>=6.0",
]

[project.scripts]
pm           = "project_registry.pm:main"
task-tracker = "project_registry.task_tracker:main"
project      = "project_registry.project_record:main"
tt           = "project_registry.time_cli:main"
deliver      = "project_registry.deliver_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"